refactor(spec-546): four modules whose names are true of their contents (t-4)
Split tool-contract.ts into the four things it actually held. Selection rule
[std-51 cl-9]: the name must be true of EVERY symbol in the module, applied
per module rather than by picking a module count.

  tool-contract.ts     436 lines, 12 exports — the shape every tool declares
                       and the context every handler receives
  doc-state.ts         102 lines,  3 exports — load a doc's state, render it
  guidance-envelope.ts 927 lines,  8 exports — one public entry point
  related-issues.ts    111 lines,  3 exports — semantic issue overlap search

FOUR, not the three dec-1 first named. related-issues is semantic search
between an Issue and the Specs or Decisions it might belong to; folding it
into guidance-envelope would make that name false of five of its symbols.
dec-1 already committed to taking the fourth file rather than keeping a small
drawer to preserve the number three — measuring the post-PR1 file is what made
it bite. dec-1, ac-9 and t-4 record the refinement.

26 exports across the four, up from 24: FooterSignal and FullDocState are
promoted because they now cross a module boundary and so have a consumer
outside their file — which is exactly when std-51 says an export may exist.

A cycle removed rather than papered over. FooterSignal declared
`issueHits: Awaited<ReturnType<typeof relatedIssuesForDecision>>`, making
tool-contract depend on related-issues while related-issues depends on
ToolCtx/ResolvedRef here. An `import type` would have compiled (types are
erased, so no runtime cycle) but left a cycle in the graph for every reader.
That derived type resolves to exactly MemexSearchHit[], so naming it directly
deletes the edge. Types only. It also fixed 8 implicit-any errors in two test
files that came from inference through the indirection.

Constraint (c) verified by measurement, not by a green suite: the confinement
window is 9314 characters — byte-identical to before the split — with
renderFooterSignal, composeSupersessionHeader and composeGuidanceEnvelope all
inside it in that order. Only the window's absolute position moved.

response-budget-funnel's caller list now reads ["agent/handlers/doc-state.ts"]
— the payoff dec-1 named: the list finally states a fact instead of pointing
at a drawer.

The suite gained exactly 6 test cases, and they are the right ones:
mutate-coverage.static-scan (std-8) and no-direct-anthropic (std-30) generate
one case per source file, so the three new modules each pick up both guards.
Enumerated with `vitest list` and diffed against t-3 rather than guessed at;
nothing was lost.

Verified: make typecheck exit 0; real .d.ts emission exit 0; zero unused
import or local in any of the four modules; full server suite 707 files /
6558 tests passed; the seven source-scanning guards run explicitly, 45 tests
passed. Three new files, all under agent/handlers/; nothing created elsewhere.
