+++ packages/server/src/agent/handlers/decisions.ts
@@ -42,11 +42,15 @@ import {
 import { requireBallotForMemex, decisionBallotTrueFacets } from "../../services/facet-ballot.js";
 import { parseBallotArg, storeRouteAndReadout, routeAndReadout } from "../../services/facet-consume.js";
 import {
-  VERBOSE_FIELD,
   formatState,
   fullDocState,
-  isDocLikeKind,
+} from "./doc-state.js";
+import {
   relatedIssuesForDecision,
+} from "./related-issues.js";
+import {
+  VERBOSE_FIELD,
+  isDocLikeKind,
   reqCtx,
   resolveRefArg,
   type ToolSpec,
