+++ packages/server/src/agent/tool-specs.ts
@@ -17,7 +17,9 @@
 // tool-manifest.ts`; the b-67 regression test asserts the two stay in lockstep.
 
 import { toolManifest } from "@memex/shared";
-import type { ToolSpec } from "./handlers/tool-contract.js";
+import type {
+  ToolSpec,
+} from "./handlers/tool-contract.js";
 
 // spec-366: re-export the shared infra symbols external modules/tests import
 // from this file, so no import site moved (std-16 contract unchanged).
@@ -36,27 +38,33 @@ import type { ToolSpec } from "./handlers/tool-contract.js";
 // while asserting NOTHING and the "one author of footer prose" invariant goes
 // unguarded, in green. Only its `sanity:` test pins the two window anchors.
 export {
-  composeGuidanceEnvelope,
-  craftActivityBlock,
-  composeStatusOverview,
-  formatAcCoverageSummary,
-  relatedIssuesForDecision,
-  relatedIssuesNudge,
-  suggestActiveSpecsForIssue,
   buildNudgeOrgBlocksGetter,
-  COMPLETION_NUDGE,
   // spec-366: re-exported because tool-specs.audit.integration.test.ts imports
   // VERBOSE_FIELD from here to assert the shared-instance identity contract.
   VERBOSE_FIELD,
   MEMEX_DESC,
 } from "./handlers/tool-contract.js";
+export {
+  composeGuidanceEnvelope,
+  craftActivityBlock,
+  composeStatusOverview,
+  formatAcCoverageSummary,
+  COMPLETION_NUDGE,
+} from "./handlers/guidance-envelope.js";
+export {
+  relatedIssuesForDecision,
+  relatedIssuesNudge,
+  suggestActiveSpecsForIssue,
+} from "./handlers/related-issues.js";
+export type {
+  StatusFacts,
+} from "./handlers/guidance-envelope.js";
 export type {
   ToolCtx,
   ToolSpec,
   ResolvedRef,
   EntityKind,
   FooterSlot,
-  StatusFacts,
 } from "./handlers/tool-contract.js";
